+++ test/Main.hs
@@ -9,9 +9,10 @@ module Main (main) where
 
 import Test.Tasty (defaultMain, testGroup)
 
+import Effectful.Tracing.CompileTest qualified as CompileTest
 import Effectful.Tracing.PropertySpec qualified as PropertySpec
 
 main :: IO ()
 main =
   defaultMain
-    (testGroup "effectful-tracing" [PropertySpec.tests])
+    (testGroup "effectful-tracing" [PropertySpec.tests, CompileTest.tests])